Black_Hole_Lab/pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "theory-zero"
version = "0.1.0"
description = "Theory Zero research OS"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
  "numpy>=1.24",
  "scipy>=1.10",
  "pandas>=2.0",
  "pyyaml>=6.0",
  "matplotlib>=3.7",
]

[project.optional-dependencies]
dev = [
  "pytest>=7.4",
  "hypothesis>=6.91",
  "ruff>=0.1.5",
  "black>=23.9",
  "isort>=5.12",
  "mypy>=1.6",
]
gpu = [
  "cupy-cuda11x>=12.0; platform_system != 'Darwin'",
]
notebooks = [
  "jupyter>=1.0",
  "ipykernel>=6.25",
]

[project.scripts]
theory-zero = "theory_zero.cli:main"

[tool.ruff]
line-length = 100

[tool.black]
line-length = 100

[tool.isort]
profile = "black"

[tool.mypy]
python_version = "3.10"
ignore_missing_imports = true

[tool.pytest.ini_options]
addopts = "-ra"
norecursedirs = ["legacy"]

[tool.setuptools.packages.find]
where = ["src"]
